cli_util: move snapshotting of the stale working copy into cli_util
This centralizes the logic so cli helpers can use it.
